ja (Japanese) translation
* Almost ja texts are in newly added files named with 'ja-' prefix.
* "routes/firstrun.js" has some ja texts in it.
* "app.js" has been modified to enable multiple language.
* A link to ja page has been added in "views/layout.jade".
* There is no ja version of "views/_modules.jade" because it is not used.
* In the ja version, 'moduleList' variable is now an associative array to hold translated module names. Some 'toUpperCase()' calls have been changed to use this traslated module names.
* The module page system in "views/ja-FRE-module-individual.jade" has slightly changed to enable the Jade like "#{var}" substition in string literals. Check the 'normal_row' mixin now has an optional argument 'd'.

We have added the language selection links in the top bar section. Since we do not know much about graphic design, please move the link texts to the right location.

The comments in sample codes in the individual module pages have not yet translated. This is because the codes come from the other Git repository. We will translate them in the future.

@sgk @asagin @mknkgw